+++ pypolyback/log.py
@@ -4,6 +4,16 @@
 import sys
 from twisted.python import log
 
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
 def start(where=sys.stdout):
     """
     Starts logging